break up big testing
we don't want our test cases to be too big to fail
